Patrolling Enemy group prefab, enemies stop
Enemies stop when you do (the ones with "Stop" in their name, and the gaggle.)
A patrolling group of 4 in case you want moving groups! Set a list of objects to patrol between for their leader, and watch them go! The gaggle will follow the leader, and the zone is centered around the leader. They will not rush you if you enter their zone. The leader also pauses when you do, but the minions do not (so they can catch up to the leader)